Add privateMethod
1. privateMethod : privateMethod 폴더를 추가하여 app.js 에서 분리
